# server/docs/swagger.yaml
definitions:
...
  loomiadmin.AnchorPointsFlows:
    properties:
      ID:
        description: 主键ID
        type: integer
      anchorUid:
        description: anchorUid字段
        type: string
      channelId:
        description: channelId字段
        type: string
      coin:
        description: 用户消费金币
        type: integer
      eventId:
        description: eventId字段
        type: string
      ext:
        description: 额外字段
        type: string
      guildId:
        description: guildId字段
        type: string
      num:
        description: num字段
        type: integer
      payUid:
        description: payUid字段
        type: string
      pointsAmount:
        description: 积分收益数
        type: integer
      pointsChg:
        description: pointsChg字段
        type: integer
      pointsChgAfter:
        description: pointsChgAfter字段
        type: integer
      pointsChgBefore:
        description: pointsChgBefore字段
        type: integer
      pointsFlowId:
        description: 积分流水id
        type: integer
      ratio:
        description: 兑换比例 10=10%
        type: integer
      source:
        description: source字段
        type: string
      sourceType:
        description: sourceType字段
        type: string
      status:
        description: status字段
        type: integer
      subSource:
        description: subSource字段
        type: string
    type: object
...
  request.IdsReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.5.9
paths:
  /anchorPointsFlows/createAnchorPointsFlows:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建anchorPointsFlows表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/loomiadmin.AnchorPointsFlows'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"创建成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建anchorPointsFlows表
      tags:
      - AnchorPointsFlows
  /anchorPointsFlows/deleteAnchorPointsFlows:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除anchorPointsFlows表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/loomiadmin.AnchorPointsFlows'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除anchorPointsFlows表
      tags:
      - AnchorPointsFlows
  /anchorPointsFlows/deleteAnchorPointsFlowsByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 批量删除anchorPointsFlows表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除anchorPointsFlows表
      tags:
      - AnchorPointsFlows
  /anchorPointsFlows/findAnchorPointsFlows:
    get:
      consumes:
      - application/json
      parameters:
      - description: 主键ID
        in: query
        name: ID
        type: integer
      - description: anchorUid字段
        in: query
        name: anchorUid
        type: string
      - description: channelId字段
        in: query
        name: channelId
        type: string
      - description: 用户消费金币
        in: query
        name: coin
        type: integer
      - description: eventId字段
        in: query
        name: eventId
        type: string
      - description: 额外字段
        in: query
        name: ext
        type: string
      - description: guildId字段
        in: query
        name: guildId
        type: string
      - description: num字段
        in: query
        name: num
        type: integer
      - description: payUid字段
        in: query
        name: payUid
        type: string
      - description: 积分收益数
        in: query
        name: pointsAmount
        type: integer
      - description: pointsChg字段
        in: query
        name: pointsChg
        type: integer
      - description: pointsChgAfter字段
        in: query
        name: pointsChgAfter
        type: integer
      - description: pointsChgBefore字段
        in: query
        name: pointsChgBefore
        type: integer
      - description: 积分流水id
        in: query
        name: pointsFlowId
        type: integer
      - description: 兑换比例 10=10%
        in: query
        name: ratio
        type: integer
      - description: source字段
        in: query
        name: source
        type: string
      - description: sourceType字段
        in: query
        name: sourceType
        type: string
      - description: status字段
        in: query
        name: status
        type: integer
      - description: subSource字段
        in: query
        name: subSource
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"查询成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用id查询anchorPointsFlows表
      tags:
      - AnchorPointsFlows
  /anchorPointsFlows/getAnchorPointsFlowsList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 主键ID
        in: query
        name: ID
        type: integer
      - description: anchorUid字段
        in: query
        name: anchorUid
        type: string
      - description: channelId字段
        in: query
        name: channelId
        type: string
      - description: 用户消费金币
        in: query
        name: coin
        type: integer
      - in: query
        name: endCreatedAt
        type: string
      - description: eventId字段
        in: query
        name: eventId
        type: string
      - description: 额外字段
        in: query
        name: ext
        type: string
      - description: guildId字段
        in: query
        name: guildId
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: num字段
        in: query
        name: num
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - description: payUid字段
        in: query
        name: payUid
        type: string
      - description: 积分收益数
        in: query
        name: pointsAmount
        type: integer
      - description: pointsChg字段
        in: query
        name: pointsChg
        type: integer
      - description: pointsChgAfter字段
        in: query
        name: pointsChgAfter
        type: integer
      - description: pointsChgBefore字段
        in: query
        name: pointsChgBefore
        type: integer
      - description: 积分流水id
        in: query
        name: pointsFlowId
        type: integer
      - description: 兑换比例 10=10%
        in: query
        name: ratio
        type: integer
      - description: source字段
        in: query
        name: source
        type: string
      - description: sourceType字段
        in: query
        name: sourceType
        type: string
      - in: query
        name: startCreatedAt
        type: string
      - description: status字段
        in: query
        name: status
        type: integer
      - description: subSource字段
        in: query
        name: subSource
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取anchorPointsFlows表列表
      tags:
      - AnchorPointsFlows
  /anchorPointsFlows/updateAnchorPointsFlows:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新anchorPointsFlows表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/loomiadmin.AnchorPointsFlows'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新anchorPointsFlows表
      tags:
      - AnchorPointsFlows
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"